requirements.txt
faster-whisper==0.10.0
torch==2.2.0
transformers==4.38.2
datasets==2.16.1
pyttsx3==2.90
gTTS==2.5.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.3
python-dotenv==1.0.0
huggingface-hub==0.20.3 